Add .gitignore, refactor imports, and update project structure
- Created .gitignore to exclude virtual environment and build files.
- Renamed EQFinder to EQ_finder for consistency.
- Updated ParameterFunction initialization in EQ_finder.
- Changed feature calculation to ensure integer type in f_X class.
- Removed setup.py and added pyproject.toml for package management.
- Updated README.md with project details and installation instructions.
- Added MIT License file.
- Cleaned up unused code in utils.py.

+++ src/audioeqinfer/densityrepr.py
@@ -9,7 +9,7 @@ from pedalboard.io import AudioFile
 
 class f_X():
     def __init__(self, sr=44100, chunk_duration=0.02):
-        features = (sr*chunk_duration)//2 +1 #because we have 44100 samples per second, and 20 milliseconds of audio, rrft size is (n/2+1)
+        features = int((sr*chunk_duration)//2) + 1 #because we have 44100 samples per second, and 20 milliseconds of audio, rrft size is (n/2+1)
         self.base_dist = distributions.StandardNormal(shape=[features])
         self.transforms = []
         for _ in range(3):
